+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of cookbook-openstack-common.
 
+## 0.4.1:
+* Changed endpoint attributes to use http for default scheme.  this is inline with
+  default settings in keystone.  fine for dev,  but should be ssl for prod.
+
 ## 0.4.0:
 * Remove `#config_by_role` as it is no longer used and no longer suits our needs.